Refactors date/date_time classes for DRYness
  - `ComparableDates` addresses the identical `Comparable` implementation
    shared between icalendar's `Date` and `DateTime` classes
  - `ParsableDates` addresses the identical parse/rescue/reraise behavior
    shared between icalendar's `Date` and `DateTime` classes
