+++ lib/icalendar/values/date_time.rb
@@ -1,10 +1,14 @@
 require 'date'
+require_relative 'comparable_dates'
+require_relative 'parsable_dates'
 require_relative 'time_with_zone'
 
 module Icalendar
   module Values
 
     class DateTime < Value
+      include ComparableDates
+      include ParsableDates
       include TimeWithZone
 
       FORMAT = '%Y%m%dT%H%M%S'
@@ -13,13 +17,7 @@ module Icalendar
         if value.is_a? String
           params['tzid'] = 'UTC' if value.end_with? 'Z'
 
-          begin
-            parsed_date = ::DateTime.strptime(value, FORMAT)
-          rescue ArgumentError => e
-            raise FormatError.new("Failed to parse \"#{value}\" - #{e.message}")
-          end
-
-          super parsed_date, params
+          parsing(value, FORMAT) { |parsed_date| super(parsed_date, params) }
         elsif value.respond_to? :to_datetime
           super value.to_datetime, params
         else
@@ -34,18 +32,6 @@ module Icalendar
           strftime FORMAT
         end
       end
-
-      def <=>(other)
-        if other.is_a?(Icalendar::Values::Date) || other.is_a?(Icalendar::Values::DateTime)
-          value_ical <=> other.value_ical
-        else
-          nil
-        end
-      end
-
-      class FormatError < ArgumentError
-      end
     end
-
   end
 end
